feat: Add Saloon and DTO framework module
The new module provides exception and response handlers for handling requests and responses. It includes usage of the `HandlesResponse` trait for any DTO and the 'pool' functionality of Saloon. Handlers have to be used when calling 'pool'. Documentation explaining the concept and usage has been included in a README file within the module.