feat: add bulk translation feature to Deepl integration
The commit includes changes in several classes related to Deepl integration. A method for handling bulk translate requests has been added to the main Deepl class. Changes were introduced to the RequestBodyData, including a method to create DTO from response & the addition of a HandlesResponse trait. Furthermore, the RequestBodyData class was renamed to TranslateTextData for clarity.